chore : IDamagable 수정
기존 : int
변경 : float

+++ Assets/Scripts/Player/Weapon.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void TakePhysicalDamage(int damageAmount);
+    void TakePhysicalDamage(float damageAmount);
 }
 
 public class Weapon : MonoBehaviour
@@ -21,7 +21,7 @@ public class Weapon : MonoBehaviour
         // 무기의 타입 검출, 무기일 경우, 무기가 가지고 있는 데미지 주기.
         //  if (damageable != null)
         //  {
-        //      damageable.TakeDamage(itemData.atk);
+        //      damageable.TakePhysicalDamage(itemData.atk);
         //  }
     }
